[finished-lesson] Use LessonCoordinator
