+++ Teacher Workout/Features/Discover/Coordinator/DiscoverCoordinator.swift
@@ -21,10 +21,9 @@ final class DiscoverCoordinator: NSObject, Coordinator {
     // MARK: - LessonIntroView Interaction
 
     func presentLessonIntroView(_ lesson: Lesson) {
-        let introView = LessonIntroView(lesson: lesson, delegate: self)
-        let viewController = UIHostingController(rootView: introView)
-        viewController.modalPresentationStyle = .fullScreen
-        navigationController.present(viewController, animated: true)
+        let coordinator = LessonCoordinator(navigationController: navigationController, lesson: lesson)
+        addChildCoordinator(coordinator)
+        coordinator.start()
     }
 }
 
@@ -34,20 +33,6 @@ extension DiscoverCoordinator: DiscoverViewDelegate {
         let viewDetails = ThemeDetails(viewModel: viewModel, delegate: self)
         let viewController = UIHostingController(rootView: viewDetails)
         navigationController.pushViewController(viewController, animated: true)
-    }
-}
-
-extension DiscoverCoordinator: LessonIntroViewDelegate {
-    func lessonIntroViewDidTapClose(_: LessonIntroView) {
-        navigationController.dismiss(animated: true)
-    }
-
-    func lessonIntroViewDidTapStartLesson(_: LessonIntroView) {
-        navigationController.dismiss(animated: true)
-    }
-
-    func lessonIntroViewDidTapSaveLesson(_: LessonIntroView) {
-        navigationController.dismiss(animated: true)
     }
 }
