+++ Teacher Workout/Features/Lessons/View/LessonFinishedView.swift
@@ -1,0 +1,7 @@
+import SwiftUI
+
+struct LessonFinishedView: View {
+    var body: some View {
+        Text("Lesson Finished")
+    }
+}
